arc: networkd: Move ARC device management into service impls.
The next step in removing device manager is to have each service manage
it's own devices. This patch starts by having the ARC implementations
deal with their own ARC device (arcbr0 or arc_br1). One side-effect of
this is that on ARCVM, the bridge is now removed when the service stops.

BUG=b:145483351
TEST=units, manual verification of play connectivity on atlas (++/VM)
TEST=arc.Boot{.vm} over and over in random order

Change-Id: Id26a1b3bb03b90598cc772d748109000fa272854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2082697
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>