cryptohome: Add encrypted reboot vault
In situations where we want to store data on disk as encrypted
but we have already requested the TPM to be cleared on the next
boot, the encrypted reboot vault acts as a transient store for
data.

Setting up the encrypted reboot vault occurs during early
boot. Once the vault is setup, users of the vault can setup and
use directories within the vault. On the next boot, if we can
find the key in pstore, we unlock the vault again; otherwise we
recreate the vault.

BUG=chromium:1025044
TEST=manual test; add vault and reboot several time.

Cq-Depend: chromium:2117108
Change-Id: I9bea9227b9b01f40ff5373cff9b0ec32cb8b9156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2069605
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>