Add /run/chromeos-config as a bind mount to minijailed services which use chromeos-config
This is a wide-change across platform2 which changes services that use
both minijail to either chroot or pivot_root, and uses libcros_config
to access model configuration.

After CL:2065478, services which require chromeos-config will need to
be able to have read access to /run/chromeos-config/v1 to read
configuration values. This adds this as a bind mount in the
appropriate upstart scripts.

BUG=chromium:1054199
TEST=tryjobs

Change-Id: I933701143e9c73f5bb663c304d040a58988ac8da
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2070325
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>