cryptohome: Increase ChallengeKey D-Bus timeout
Increase the timeout of the "ChallengeKey" D-Bus call from cryptohomed
to Chrome from 25 seconds to 2 minutes.

This makes the UX slightly better (as the user has more time to enter
the smart card PIN during the login, which is one of the things that
the "ChallengeKey" is doing internally).

To note, the new timeout matches the timeout of the "MountEx" and
"CheckKeyEx" D-Bus calls from Chrome to cryptohomed, which is
technically not required but makes sense, as the "ChallengeKey" call is
nested into "MountEx" and "CheckKeyEx".

BUG=chromium:1051866
TEST=start logging in as an existing user via a smart card, don't enter any PIN, verify that the login attempt gets aborted only after ~2 minutes

Change-Id: I983814c4b619d5b4e5b0897f81a31029f0c1a7ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2075397
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>