chromeos-config: fail when paths do not start with "/"
The ChromeOS ConfigFS implementation of this method is lenient when
slash is not passed in the front of a config path. Lets complain when
others try to do this, as apparently there is already some confusion
around this (see CL:2082701).

There was a unit test for this, but was not failing as unit tests use
JSON config at the moment instead of the FS view. crbug.com/1057661 is
open for this.

BUG=chromium:1054199
TEST='cros_config "" name' and 'cros_config firmware image-name' fail

Change-Id: Ie1123a1ddfe8aaf9b2382cb757cd158a0536e0cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2083492
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>
Commit-Queue: Yuval Peress <peress@chromium.org>