runtime_probe: add /run/chromeos-config/v1 as a bind mount to minijail
This is a wide-change across platform2 which changes services that use
both minijail to either chroot or pivot_root, and uses libcros_config
to access model configuration.

After CL:2065478, services which require chromeos-config will need to
be able to have read access to /run/chromeos-config/v1 to read
configuration values. This adds this as a bind mount in the
appropriate upstart scripts.

BUG=b:150658730
TEST=/usr/bin/tast run -varsfile="$VARSFILE" "$ACTIVE_DUT" \
     '(!disabled && "group:runtime_probe")'

Change-Id: Iafdd570977e77016bfbfc33dde6be2fcef1e7fa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2084652
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>