shill: Make RpcIdentifier a dbus::ObjectPath
Semantically, the only difference between a std::string and an
RpcIdentifier is that an RpcIdentifier is assumed to be something that
one could get from dbus::ObjectPath::value(). Since RpcIdentifier must
be a different type from std::string in order for templatization of
KeyValueStore and PropertyStore to work cleanly, making it an alias to
dbus::ObjectPath seems like a logical choice.

This is essentially a reland of CL:1646937, but was done from
scratch to ensure invalid references are not being passed around and
because the codebase has changed enough to make merge resolution as
much work as redoing this.

BUG=chromium:936111
TEST=All unit tests are passing.

Change-Id: I0d3b3a70838f57ac421e4f402936126a32e70840
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2086279
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>