arc: networkd: Invert virtual interface creation.
This patch flips the logic that creates the vritual interface pair for
ARC and does so inside the container net namespace before pulling the
bridged end back into the default. A change in Shill caused it to start
paying attention to the peer end. This was causing log spam (at best)
and potentially race issues causing the interface to go down (at worst).

BUG=b:148416701
BUG=b:149311615
TEST=units & manual on atlas, nocturne
TEST=arc.Downloads,arc.Reboot,arc.Boot,
TEST=arc.MultiNetworking,arc.MulticastForwarder,arc.IPv6Connectivity

Change-Id: I274381211314cda19344b81dc5e129cb8d1363ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2087191
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>