vm_tools: reset IPv6 in termina VM upon default network change
There is a bug in current implementation of IPv6 for Crostini with
NDProxy that IPv6 configurations in Crostini are not updated when a
network change happens. This patch lets concierge listen to shill
signals and trigger an IPv6 stack reset upon default network change.
This will force a new round of SLAAC process on new network and get
the correct configuration.

A follow up patch will be needed to apply the similar change in lxd
container to fully address the issue.

BUG=chromium:1050165
TEST=manual: deployed on atlas, verify ipv6 reconfigured upon
     network change.

Change-Id: I52fe3ba9229be5a138f16be073d05727a125d345
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2087193
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>