libbrillo: fail hard if SecureAllocator fails to allocate memory
In crbug.com/1056318, we see that SecureAllocator return nullptr
on failing to allocate memory is a very bad idea. The std::allocator
is expected to return std::bad_alloc on failing allocations and hence
returning a value is taken as an indicator of a successful allocation
by STL. Then, we go on to fail at the first spot the purported
allocated memory is accessed.

Since exceptions are disabled, we can't throw std::bad_alloc. Instead,
fail hard in SecureAllocator directly: this is easier to correlate and
allows developers to understand that this was a specific failure.

BUG=chromium:1056318
TEST=unittests
TEST=unittests after commenting out the ScopedClosure.Release() result
     in a segmentation fault.

Change-Id: Ic7dd4e8ce233c2996a375627a1f19394dc340bcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2090777
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>