arc: networkd: Remove device manager.
This patch removes the remaining use of device manager and moves the
logic that manages shill devices for ARC into the arc service. Now these
devices are deleted when the guest disconnects and recreated on
reconnect - including the ARCVM bridge but excluding the ARC++ bridge.

BUG=b:145483351
TEST=manual on atlas and nocturne
TEST=various tasts for both container and vm

Change-Id: If68ebfb6376a85d90ed800a65d974b311dddd8fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2094401
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>