attestation: PostStartCertificateTaskOrReturn
Previously w/o extracting the introduced help function the program flow
is very confusing so that we have to write lots of comments and still
struggle knowing  why the code is like this. W/ with CL, now the code
is more self-explanaroty and potentially we can reuse it in the future.

Thus, the logic deserve to be a helper function.

BUG=b:149723745
TEST=build ok

Change-Id: Icd0d7656a3e759e5cf00bce91d12267e07db9741
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2103972
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Louis Collard <louiscollard@chromium.org>