camera: Add retry mechanism when connecting to JEA/JDA
Since we might run camera.LibJEA test after camera.CaptureUnittests,
sometimes camera.LibJEA will start when the camera dispatcher is not
ready. As a result, we might need to retry a few times if the connection
is not built successfully.

And so does JDA.

BUG=b:149889552
TEST=tast run [DUT] camera.CaptureUnittests camera.LibJEA (on scarlet)

Change-Id: If267daadc5574a7176447dd81c48b00867b63779
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2095650
Tested-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>