init: clobber: split UI into separate class
Create a new class ClobberUi, which is responsible for printing progress
to the terminal during clobbering.

The main thread reports progress to ClobberUi, and the UI thread wakes
up periodically to update the screen. It is also responsible for
managing the file handle to the terminal, so that only one UI is being
displayed at a time.

BUG=chromium:1045178
TEST=unit tests, clobber on-device

Change-Id: Ief9e6f383b55420bba4ad545cc0798a60bd886f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2097100
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>