crash: Don't log the crashing executable name inside VMs
For VM applications, the name of the process that crashed can be
sensitive information. Since the VM logs currently go to
/var/log/messages on the host, which is accessible to any user on the
device, we should not log these names there.

This CL makes a short term, safe fix by just not logging certain
messages inside the VM. This may not be comprehensive on the error
paths, but should at least prevent data from leaking in the normal
case.

In the long term we will want to move the VM logs somewhere else and
re-enable these messages.

BUG=chromium:1053847
TEST=Crashed a program inside the crostini VM and checked /var/log/messages

Change-Id: If98932d7199e024331a2816e6c05cdac1c444728
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2096294
Tested-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>