arc: Handle partial write correctly
If the socket type is SOCK_DGRAM or SOCK_SEQPACKET, sendmsg/write simply
returns EAGAIN when the data cannot be fit into the socket buffer.

OTOH if the socket type is SOCK_STREAM, sendmsg/write returns EAGAIN
only when the socket buffer is full.
If the data cannot be fit in the socket buffer but it's not full,
it sends part of data.

Rewrite tests to verify this behavior.

BUG=b:151287452
TEST=cros_workon_make --test arcvm-vsock-proxy

Change-Id: I733b1760c5687ef0323e7cb4cf7e9109fcc08740
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2101052
Tested-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>