arc: networkd: Make ArcVmImpl P/R aware.
This first patch adds a "feature" to the arc service that enables it to
be use (or not) multinet for arcvm. This is disabled and there is no
functional change but it sets the stage for the next patch(es) that will
take advantage of this.

BUG=b:136417971
TEST=units, tast arc.PlayStore.vm, android-sh ping google.com

Change-Id: I41fff3fd7b5e2723b7f0c24e75d27e0967cc0b44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2105177
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>