diagnostics: add --force_cancel_at_percent flag to diag
Add a new boolean flag to diag, --force_cancel_at_percent,
which runs a routine and tries to cancel it after a certain
percent. This flag will be useful in debugging problems with
cancelling routines, and will allow us to add routine
cancellation to the diagnostics tast test and prevent
regressions in cancellation behavior in the future.

BUG=chromium:1061163
TEST=diag --action=run_routine --routine=urandom
--force_cancel_at_percent=20

Change-Id: I5f0625bb5ebe12ea679741402bb458aa3408a858
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2105879
Reviewed-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>