diagnostics: move RunRoutineAndProcessResult to DiagActions
Having RunRoutineAndProcessResult as a class member allows
it to take advantage of the DiagActions' state (the mojo
adapter, the id of the routine) without explicitly needing
to pass them around. This will help in the future when
RunRoutineAndProcessResult is split into smaller functions,
and each of those functions doesn't need to be passed an
id and adapter.

BUG=chromium:1061163
TEST=emerge-nami diagnostics

Change-Id: I25148f35ee1d245efe154064a2c71d0cdd25ef45
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2106933
Reviewed-by: Jesse Schettler <jschettler@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>