init: Add test script to keep crash_reporter state
We use two flag files under /run, crash-test-in-progress and
mock-consent, to signal to the crash reporter that it should use
different behavior for tests. For most tests that's fine, but for tests
that reboot the machine, these files are nuked on reboot. Add a test
service to restore these files from /var on reboots (for test images
only).

BUG=chromium:1060344
TEST=Deploy to an eve, touch /var/spool/crash-test-in-progress and \
     mock-consent, reboot, check /run/crash_reporter

Change-Id: I789b98506bb2d7c807eabf79a612af15b221a5f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2107802
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>