login: Start breakpad/crashpad A/B experiment
Start an experiment where 10% of Chrome sessions are started with
--enable-crashpad. We can then check that approximately 10% of all
Chrome crashes (from participating machines) are from crashpad, to prove
that crashpad is capturing the same fraction of crashes as breakpad.

This isn't a finch experiment because finch is initialized after
breakpad / crashpad. So the experiment has to be in session_manager.

Some boards are known to only work with crashpad, and some boards are
known to only work with breakpad. Exclude those from the experiment.

BUG=chromium:1038427, chromium:1051354
TEST=Ran on veyron_minnie; saw that occasionally Chrome would be started with --enable-crashpad
TEST=Ran on hana; after 20 login/logout cycles and 10 restart uis, Chrome was always started with --no-enable-crashpad
TEST=Ran on hana-kernelnext; after 5 login/logout cycles and 5 restart uis, Chrome was always started with --enable-crashpad

Cq-Depend: chromium:2110540
Change-Id: I51b0935cc6bf5a4fcd94aa119ebd4ebb88047b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2110751
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>