chromeos-config: Re-enable converter test
The converter depends on the config repo for proto (API) bindings.

Previously, this required cros-config-api (an sdk dep) in order to use
the bindings, which is unnecessary and makes API changes and local
development more difficult.

This leverages the cros_workon architecture and portage to setup the
local src tree correctly, so the config repo and bindings are
consistently available in the local dev, chroot dev, and portage build
environments.

BUG=None
TEST=via portage emerge, run_tests in the chroot, and outside the chroot

Cq-Depend: chromium:2109993
Change-Id: I1df9d130d6ecd47510b59dd634afd89e55ee1000
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2109501
Tested-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>