crash: Write crash_reporter errors to separate log
Previously, if crash reporter had an error during collection, it would
write information about that error to a crash using the same exec_name
as whatever binary it ran into trouble with (and a basename based on
that binary). This is suboptimal for two reasons:

1) It makes it hard to debug -- the team who owns that binary gets a
report that is useless (to them), and the telemetry team has a hard time
of finding the relevant error reports.
2) It can cause spurious failures. For example, if the crashed binary
generates a log file, then crash_reporter will try to use the same
filename, but fail because a log file with that name already exists.

BUG=chromium:1063103
TEST=test_that --board=eve $DUT logging_UserCrash
TEST=tast run $DUT platform.UserCrash.*

Fixes: chromium:1063103
Cq-Depend: chromium:2138591, chromium:2136203
Change-Id: I88573ef49e983944fb3ee3e951cff2f762cd8732
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2136289
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>