shill: determine regulatory domain at startup
If Shill starts after the initial regulatory domain is determined, we
won't log any metric initially. We *may* log one later (e.g., when
scanning or connecting, and the driver updates its domain), but that's
not guaranteed; for instance, if the chip has a country-specific EEPROM.

To resolve this inconsistency, probe for the current regulatory domain
at device startup. We ask for the wiphy-specific domain, which should be
the source of truth (e.g., with "self-managed" regulatory like in
iwlwifi); if the wiphy does not provide its own domain, the kernel will
return the system-wide domain instead.

We also may want to retrieve attributes like NL80211_ATTR_DFS_REGION,
which are not available in CHANGE messages -- only GET. So instead of
doing a complex mixture of CHANGE and GET logic, let's just request more
info (GET_REG) for every CHANGE event.

BUG=b:150971082
TEST=`tail -f /var/log/net.log | grep -i 'shill.*reg' &
      start shill SHILL_LOG_LEVEL=-7 SHILL_LOG_SCOPES=wifi`, with and
      without a pre-existing (non-00) regulatory domain; tested on
      rtw88, with and without this patch:
        https://lkml.kernel.org/r/20200313034918.22222-2-yhchuang@realtek.com
      and tested on ath10k
TEST=unit tests

Change-Id: I8e185af02de115d9ac59da77896d9a1cd080b9c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2111743
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>