shill: drop LOG from netlink MATCHERs
Gmock really doesn't like it (read: gets itself into a deadlock) when
you're running mocked functions from within a matcher. It just so
happens that the WiFiMainTest.ScanRejected test (among others) runs a
mock log instance (yuck!) while also matching Nl80211 commands. If we
ever hit any of these logging scenarios in such a test...we deadlock.
That was fun to debug.

Let's just remove the logging in the matcher, as it's not that helpful.

BUG=none
TEST=unit tests, with subsequent CL

Change-Id: Ia8d5149719c19ad1dd60da639e67cd91e14a2b97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2111742
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>