diagnostics: give cros_healthd access to kill
cros_healthd is crashing when it attempts to cancel the urandom
routine. The cancellation needs access to the kill syscall, which
isn't in cros_healthd's seccomp filter. Add it :)

BUG=chromium:1060627
TEST=diag --action=run_routine --routine=urandom --force_cancel
Note that this test requires crrev.com/c/2105879, although the
change is not dependent on that change.

Change-Id: I414267eee0f150e728a1165949b5da578734bbbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2106683
Reviewed-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>