kerberos: Add UMA for Kerberos encryption types
This histogram will collect the encryption types used when a Kerberos
TGT is successfully created or refreshed. The values can be 'all',
'legacy' or 'strong'.

Encryption types are set through Kerberos configuration in three
different fields. The values of these fields are combined and
mapped into one of the possible values for this metric.

BUG=chromium:1062630
TEST=cros_workon_make --board=amd64-generic --test kerberos

Change-Id: Ieaf38d49827e1e746844723b0d8e10b1d8414f3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2117121
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>