permission_broker: Add logging to debug claimed USB interface deny
This change adds logging to DenyClaimedUsbDeviceRule to debug the reason
why the interface caused a DENY response. Most importantly this logs the
name of the driver with a claim on the device so that we can see whether
there is another process or driver fighting for control of the device.

BUG=b:152070161
TEST=Ran unit tests.

Change-Id: Ib17cb37c48949168c0eb44e43dd90ea854f1ee86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2118615
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Tested-by: Reilly Grant <reillyg@chromium.org>