shill: fix a (potential?) dangling reference
in crrev.com/c/2116643, briannorris@ noted that this code has the
potential to exhibit the same problem as the code fixed by said CL.

At best, this CL fixes a bug. At worst, it's a nop that fixes a dangling
reference if anyone calls `dns_servers()` after
`ExpectPingDNSSeversStart`, but before installing a new mock of that
function.

BUG=None
TEST=CQ

Change-Id: I116aceff155400266e0b229043db096f1fbd4624
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2118653
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>