shill: Move RoutingTable::ConfigureRoutes to Connection
RoutingTable is a simple abstraction providing basic routing
functionality. In particular, it should not know about IPConfig, which
serves as a bridge between Device/Service and
Connection. Given this and the fact that RoutingTable::ConfigureRoutes
is used exclusively for Connection, it makes sense for it to be a
Connection method. In addition, moving this to Connection increases
code coverage of the Connection unit tests.

CL:2118994 extends this method so that included routes can also
influence routing policy.

BUG=b:152322923
TEST=All unit tests are passing.

Change-Id: I8bd2010a50df8ecb9f9fc753bea768414c1e12fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2118993
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>