camera: hal: intel: Remove stream swapping in IPU3CameraHw
There was a restriction that if there are more than two video streams.
The video0 must be continuously, but video1 doesn't follow the rule, so
there is a logic swapping the first HAL_PIXEL_FORMAT_IMPLEMENTATION_DEFINED
stream to the second place. According to intel, the restriction has
gone in current kernel, so removes the logic.

BUG=b:150254366
TEST=android.hardware.camera2.cts.RecordingTest

Change-Id: I5a805d01cff3090b86ed01a8d032c80884c63bdb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2120224
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Liang L Yang <liang.l.yang@intel.com>
Commit-Queue: Hanlin Chen <hanlinchen@google.com>