arc: find and generate key as CKO_SECRET_KEY
We generate a key with C_GenerateKey and set CKO_DATA as object class,
but that does not work. Chaps stores the key as CKO_SECRET_KEY instead
and the initial key can't be found.

This change fixes it by generating and finding the key as
CKO_SECRET_KEY.

BUG=b:139142065
TEST=manual - install a demo app that generates a key
            - log out and log back in
            - check the demo app can still use the key

Change-Id: Id77eb9b07c107f6a2bd0812ae4f23f00771ded0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2120359
Tested-by: Edman Anjos <edman@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Edman Anjos <edman@chromium.org>