trunks: Reland "add ECC salting key and sessions support"
Supports ECC salting key in trunks. Implements SW ECDH channel secret
creation in the session manager.

This CL is a reland of the second half of CL:2069983. I am splitting
the original CL so that it's easier to review. Check diff between
patchset 1 and the latest patchset to see the changes.

Differences between the 1st and 2nd attempts:
1. adding a sanity check for Z point not being the infinity point,
2. adding retry logic for the ECDH key generation and computation.

BUG=chromium:879797
BUG=chromium:910495
TEST=unittests;
     tested along with CL:2121818 in remote tryjobs (nami, octopus,
     grunt) and checked if
       1. TakeOwnership no longer bad auth errors
       2. cryptohome TPM live tests no longer had TPM_RC_REFERENCE_H0
          (used to be very flaky with ECC salted sessions)

Cq-Depend: chromium:2121818
Change-Id: I47df9e1104984b774cfbbf7d15dc45d989b31597
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2123701
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Commit-Queue: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>