init: Mount cpuset in v2 mode if possible
This is a reland of the same CL with bash quoting issues fixed.

On systems with cpuset_v2_mode, we should use it to get cgroup v2 cpuset
behavior in cgroup v1 so that the kernel does not need cpuset hotplug
fixes to v1 and makes backporting efforts much less of ancient kernel
patches (both in terms of forgetting to backport, and backporting
incorrectly).

The difference between v1 vs v2 is:
1. Offlining all cpus in a set don't migrate tasks in v1.
2. Offlining and onlining cpus restores CPUs in v2.
3. Differences in behavior of having child be subset of parent's CPUs.

For ChromeOS, we want the behavior in #2. #1 and #3. don't effect us. We
can make v1 cpusets behave like v2 with the cpuset_v2_mode mount option
which fixes the hotplug issue.

BUG=b:152062524
TEST=Boot with both types of systems and verify cpuset mount.

Change-Id: I28a1f545380ea0fe7e66bcc5fbdd26c1f9b6dde6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2134507
Tested-by: Joel Fernandes <joelaf@google.com>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Auto-Submit: Joel Fernandes <joelaf@google.com>