vm_tools: Allow setting metrics consent for testing
The crash reporter tests use a mock consent file which causes crash
reporter to ignore the actual metrics consent state. But because the
mock file is not accessible from the VM, the in-VM crash_reporter will
not produce crash reports. Make cicerone check for this when telling
the VM whether or not metrics consent has been granted on test images.

BUG=chromium:703926
TEST=tast run $IP_ADDR crostini.CrashReporter.download

Change-Id: I3d44650def3cb776620a12c6e42888838c7b32a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2123699
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>