shill: service: Resort Services on Priority/ManagedCredentials update
The Priority and ManagedCredentials Service properties can have an
impact on the sorting order of Services. In the case of multiple
already-connected Services, we can't rely on state changes to
trigger sorting, so we manually trigger it when these properties are
changed on a connected Service.

BUG=chromium:1064362
TEST=All unit tests are passing.
TEST=Manual testing by:
  * Modifying Service::Compare to ignore technology
  * On a device with both wifi and ethernet, successfully change the
    order of the connected Services by changing Priority values.

Change-Id: Ie8de78847823f50e044651e57bde5621d1ad6780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2124780
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>