vm_tools: Allow forwarding logs to a file.
vm_concierge:
Invocations of crosvm for a termina vm have "serial" set to a
file in the user's cryptohome.

syslog:
If vmlog_forarder has log_destination set to /dev/log, logs
are forwarded via unix domain socket to the destination.

If log_destination is not /dev/log, the arg is interpreted
as a filepath. If the path starts with "cryptohome" the
filepath is inside the primary user's cryptohome. For
this to work, vmlog_forwarder must run as user "crosvm"
instead of user "syslog".

login_manager:
Allow crosvm to query the primary user session.

This enables vmlog_forwarder to run as crosvm and find the
correct cryptohome in which to write logs.

BUG=chromium:1053847
TEST=deployed login_manager, vm_host_tools to DUT, verified
logs written into cryptohome.

Change-Id: I53a0b42a66f3119b0b6df707879553961128fb4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2124346
Tested-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Auto-Submit: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>