smbfs: Implement recursive delete in smbfs
Adds a method to the SmbFs Mojo interface that allows a directory to be
recursively deleted without the overhead of multiple FUSE calls (ie. the
client application will no longer need to ReadDir() and Unlink()
repeatedly.)

Instead, a single call to DeleteRecursively() will begin a series of SMB
calls to recursively traverse and delete of all descendants of the
directory and finally the directory itself.

BUG=chromium:1064471
TEST=cros_run_unit_tests --board=nami --packages smbfs
TEST=Manually tested by creating a directory containing files and
subdirectories and then deleting it via Files.app

Change-Id: Icf08b726057d7a594b93cc3f21e94b0b620bf783
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2127046
Tested-by: Josh Simmons <simmonsjosh@google.com>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>
Reviewed-by: Anand Mistry <amistry@chromium.org>