cryptohome: Add a profile to request Jetstream certificates
This allows manual testing for Jetstream devices when new ICAs are
introduced for new model names.

Test with the following commands on a Chrome OS test network (or
adjust flags and use https://chromeos-ca.gstatic.com for the PCA
for prod once this is landed):

  # cryptohome --attestation_server=test \
      --action=tpm_attestation_start_cert_request --profile=jetstream \
      --output=/tmp/cert_request
  # curl -s -H "Content-Type: application/octet-stream" \
      --data-binary @/tmp/cert_request -o /tmp/cert_response \
      https://asbestos-qa.corp.google.com/sign
  # cryptohome --attestation_server=test \
      --action=tpm_attestation_finish_cert_request \
      --name=attest-jetstream \
      --input=/tmp/cert_response --output=/tmp/cert_result
  # openssl x509 -in /tmp/cert_result -text -noout | \
      sed -n -e 's/.*Issuer: CN=\([A-Z][A-Z]* Vorlon ICA\).*/\1/p'
  DEFAULT Vorlon ICA
  #

Note that this may fail on non-Jetstream devices after we change the PCA
to not issue Jetstream certificates to unrecognized devices. In this
case, simply test on a Jetstream device.

BUG=chromium:1068235
TEST=See above

Change-Id: Ia5454d1c0482b5aa0ad53a93ed904e6d9effc879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2136894
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Auto-Submit: Yves Arrouye <drcrash@chromium.org>