debugd: Remove PLOG(WARNING) from cpuset update.
Failing to update the cpuset file is harmless on patched kernels. But
debugd sees the PLOG(WARNING) and incorrectly thinks the entire
operation failed. Because we need a low risk fix to unblock the HT
work, dropping the warning is a preferred workaround to re-writing the
error parsing.

BUG=b:152026859
TEST=FEATURES=test emerge-${BOARD} debugd; deploy; test

Change-Id: I9b13133b2330e216b11f6cb9a89c0153ea4d693c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2127975
Tested-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>