dlcservice: Reorganize dlc_manager.cc
The DlcManager had a PIMPL design that would make it hard to review the
code also it was doing many things. This CL breaks the DlcManager into
another class DlcBase. DlcBase is a class that manages the lifetime of a
single DLC (in contrast to DlcManager which was managing all DLCs). This
simply allows us to have different types of DLC implementations and
extract that logic out of DlcManager. In this CL, DlcManager only acts
as a loop over the list of supported DLCs for each requested
operation.

The changes to DlcService and unittests are minimal to ensure we did not
change the core logic of dlcservice.

One of the big changes in this new design is that we won't constantly
hit the disk for getting information we need like reading manifest or
getting list of packages, etc. For all supported DLCs we load the
manifest and other properties into the instance of the DlcBase class and
we continue to rely on these information only. This improves the
dlcservice performance.

BUG=chromium:1042987
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I85c48405814403aba2b64d9a9bb32e2bf7905469
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2128950
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>