authpolicy: Refactor chromad encoding policies
Chromad encoding policies were using a forbidden functions
such as std::function, and it was needed to migrate that
into using callback instead.
This CL is doing the following:
- Migrating the usage of std::function into callbacks
- Using a common PolicyOptions to be set for any policy, which
supports that, between user and device policies.
- Making the creation of the proto along with its PolicyOptions
being done after retrieving the policy value, instead of having
it as a std::function or base::callback

BUG=chromium:1056218, chromium:1053329
TEST=cros_run_unit_tests --board=amd64-generic --packages authpolicy

Change-Id: I1a8acb1f85315f77c332a07b3f6998ddee6a3c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2130251
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Tomasz Dobrowolski <tomdobro@chromium.org>
Tested-by: Amr Aboelkher <amraboelkher@google.com>
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>